Rename h.badge -> h.blocklist
