+++ h/app.py
@@ -62,7 +62,7 @@ def includeme(config):
     config.include('h.accounts')
     config.include('h.admin')
     config.include('h.auth')
-    config.include('h.badge')
+    config.include('h.blocklist')
     config.include('h.claim')
     config.include('h.groups')
     config.include('h.notification')
