+++ h/blocklist/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-from h.badge.models import Blocklist
+from h.blocklist.models import Blocklist
 
 
 def includeme(config):
